修复一些问题

+++ BiliAPI/BiliDynamic/DynamicModel/BiliDynamicAttentions.cs
@@ -2,6 +2,6 @@
 {
     public struct BiliDynamicAttentions
     {
-        public int[] uids { get; set; }
+        public long[] uids { get; set; }
     }
 }
