+++ BiliAPI/BiliDynamic/DynamicModel/BiliDynamicOrigin.cs
@@ -5,7 +5,7 @@
         public long? uid { get; set; }
         public int? type { get; set; }
         public long? rid { get; set; }
-        public int? acl { get; set; }
+        public long? acl { get; set; }
         public int? view { get; set; }
         public int? repost { get; set; }
         public long? dynamic_id { get; set; }
